requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
Pillow>=9.5.0
notebook>=7.0.0
ipykernel>=6.0.0
